Drop unnecessary `wheel` from build deps
This was never needed, but some docs used to show it in examples in the past. I corrected that in the `setuptools`' docs a while back. Older `setuptools` used to depend on this project via a PEP 517 hook when building wheels, while the modern versions vendor it now (since a little over a year ago).